Price計算でintにできる部分はintにした

+++ vendingmachine/src/vendingmachine/VendingMachine.java
@@ -102,7 +102,7 @@ public class VendingMachine {
 	}
 
 	private void makeChange() {
-		int totalChange = insertedCoin.totalAmount() - selectedProduct.price().intValue();
+		int totalChange = insertedCoin.totalAmount() - selectedProduct.price();
 		Coin changeCoin = CoinAssorter.changeCoin(totalChange);
 		if (changeCoin != null) {
 			change.add(changeCoin);
